Don't automatically name new pads based on search query
We observed that users frequently created new pads with titles that look
a lot like search queries.  This pollutes workspaces with empty pads.
Worse, these pads have titles optimized for search ranking since they
are named after popular queries, so search results quickly become fairly
useless.

We're not sure exactly what flow leads users to create these pads, but
we suspect that they're clicking the "new pad" button on the search
results page, which fills in a `title` query param based on the search
query.  This commit removes the auto-population of the `title` so the
new pad will be given the glorious default title of "Untitled."