[install] Update default values in bin/exports.sh
We also clean up exports.sh a bit so it's easy to tell what needs
customizing for new environments vs. what is supposed to work as-is.
These are hopefully more up-to-date and reasonable defaults for a
developer on OS X, and for developers not on OS X, hopefully the cleanup
helps clarify what likely needs customization.

One of the changes to default values is to use the scala-library.jar
within the given installation of Scala, not the repo's
scala-library.jar.  I don't know much about Scala, but it seems like a
better default to make SCALA, SCALA_HOME, and SCALA_LIBRARY_JAR all
match instead of SCALA_LIBRARY_JAR being a special snowflake that's
checked into the repo.