Remove 'create pad' row from search autocomplete
While this is intended to be a feature, I suspect people select it
accidentally more often than they do intentionally.  Selecting this row
issues a request to `/ep/pad/newpad?title=QUERY&r=POSITION`, where QUERY
is the query in the search bar and POSITION is the position of the row.
We saw this kind of URL in the request log for a user who accidentally
created a pad.  And anecdotally, I've personally selected this row by
accident.

Even if we remove this feature, it's relatively easy to create a new pad
by clicking the New Pad button, so hopefully we're not taking too much
away from users.